feat (router-storage): Introduce state management (Pinia) and Vue router for login/register
I scraped the previous code and made the login and register page as seperate component, which was then feeded to the Vue application as dependencies.
I struggled for some time with how to make components fit inside the same script tag, as well as to properly get the boundaries between Options API and Composition API. With the help chatGPT I have learned more, but there are still some unknows elements to it (which I will probably learn while using Vue JS more)

The last item that took alot of my time was integrating Pinia to the app. On the face of it, it turned out to be quite simple, but there were so many confusing stuff me. The main reasons being that Pinia creators expect you to use Composition API. Although they were also prividing Options API code, this was even more confusing since Composition API can in some way be writen as an Otions API.

For now I will halt the development on this feature to further focus on the backend that will handle the authentication.