feat (UI): Design ToDo component UI
This one took a lot of time because I had to learn Figma and then design it on there, before I could implement it.

Moreover, while design the icon button, I first used a normal icon and then added interactivity on it. However I was more and more recreating the functionality of button. From this conclusion, I scraped the whole thing and used a button as the parent of the icon (button.button span.icon)

At the end, that was the best decision since implementing all the functionalities of a button is time consuming and require good knowledge of design, something I lack thoroughly.