feat: Fetch all todos from server and display them to the client (with Vue JS)
This one was tricky since a bug stoped me for quite some time.

It turned out that the issue was caused by this 'localhost:2200/api/v1/todo', more precisely by the absence of 'http://' in front of that url string. That's it ! It took me so much time to figure it out. It was indeed by chance, with the help of ChatGPT, that I found the solution.